Remove removed? from doc-upgrader; call in same fashion through code
the removed comes from the (.-type ...) of the
https://firebase.google.com/docs/reference/js/firebase.firestore.DocumentChange

This is type can be "add", "remove", or "update". This just is not something
I am choosing to expose in firemore. I think it makes more sense to always
return the query as a ordered vector, not just the documents that have modified
within the query.