Now promise->chan has success and auto logging failure.
on-success should be a (ideally pure) function that takes the value returned
from the promise and does some cleanup on it. on-failure is ideally pure,
returns a value that will be logged and returned as a ex-info, and can be used
as a side effecting function with no logging by returning nil.